combine products and proxies
